requirements.txt
asgiref==3.5.0
Django==3.2.10
django-bootstrap-form==3.4
djangorestframework==3.12.4
Pillow==8.2.0
python-dateutil==2.8.2
pytz==2021.3
six==1.15.0
sqlparse==0.4.1
whitenoise==5.3.0
django-tinymce==3.4.0
